perf: replace O(n²) event scanning with sweep-line active-event tracker
The render loop in pipeline.rs checked every event against every
timestamp (events.iter().any(...)), producing O(n×m) complexity
where n=events and m=timestamps. For a 2h movie at 24fps with
~2000 events, this is ~345M checks.

Replace with a sweep-line approach: sort events by start time once,
then maintain a BinaryHeap (min-heap via negated values) of active
event end times. For each timestamp, add newly started events and
remove expired ones. Both operations are O(log n) per timestamp,
reducing overall complexity to O(n log n + m log n).

+++ crates/subtitle-renderer-libass/src/domain/composer.rs
@@ -28,10 +28,10 @@ pub fn compose_frame(images: &[AssImageData], width: u32, height: u32) -> RgbaFr
 
         // libass stores color as 0xRRGGBBAA (MSB = R, LSB = alpha in ASS convention).
         // ASS alpha: 0=opaque, 255=transparent. RGBA alpha: 0=transparent, 255=opaque.
-        let cr = ((img.color >> 24) & 0xFF) as u32;
-        let cg = ((img.color >> 16) & 0xFF) as u32;
-        let cb = ((img.color >> 8) & 0xFF) as u32;
-        let color_alpha = 255 - (img.color & 0xFF) as u32;
+        let cr = (img.color >> 24) & 0xFF;
+        let cg = (img.color >> 16) & 0xFF;
+        let cb = (img.color >> 8) & 0xFF;
+        let color_alpha = 255 - (img.color & 0xFF);
 
         // SIMD constants for Porter-Duff blending
         let src_r = u32x4::splat(cr);
